requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
colorama==0.4.6
pymongo==4.6.1
bcrypt==4.1.2
flask-mail==0.9.1
reportlab==4.0.7
python-dotenv==1.0.0